+++ redis_views/response.py
@@ -5,20 +5,23 @@ from django.template.base import Template
 from django.conf import settings
 from django.http import Http404
 
-import redis
-
-logger = logging.getLogger(__name__)
+from  redis import StrictRedis
 
 
 class RedisTemplateResponse(TemplateResponse):
 
+    REDIS_CLIENT_SOCKET_TIMEOUT = 2
+
     def __init__(self, *args, **kwargs):
         super(RedisTemplateResponse, self).__init__(*args, **kwargs)
-        socket_timeout = 2
+
+        self.logger = logging.getLogger(self.__class__.__name__)
+
         if hasattr(settings, 'REDIS_CLIENT_SOCKET_TIMEOUT'):
-            socket_timeout = settings.REDIS_CLIENT_SOCKET_TIMEOUT
-        self.connection = redis.StrictRedis.from_url(settings.REDIS_URL,
-                                                     socket_timeout=socket_timeout)
+            self.REDIS_CLIENT_SOCKET_TIMEOUT = settings.REDIS_CLIENT_SOCKET_TIMEOUT
+
+        self.connection = StrictRedis.from_url(settings.REDIS_URL,
+                                               socket_timeout=self.REDIS_CLIENT_SOCKET_TIMEOUT)
 
         # ping the server to make sure we have a valid connection. This will raise a
         # ConnectionError if the redis URL is invalid.
@@ -31,9 +34,8 @@ class RedisTemplateResponse(TemplateResponse):
 
         template_body = self.connection.get(template)
         if not template_body:
-            # log an error message if the template key was not found in the DB
-            msg = 'Reids Template: Could not find the template key {}'.format(template)
-            logger.error(msg)
+            error = 'Could not find the template key {}'.format(template)
+            self.logger.error(error)
 
             raise Http404()