+++ tests/test_response.py
@@ -1,17 +1,20 @@
 import pytest
 import mock
 import redis
+import logging
 
 from django.http import Http404
-from redis_views.response import RedisTemplateResponse, logger
+from redis_views.response import RedisTemplateResponse
 
 
 class TestRedisTemplateResponse:
 
     @pytest.fixture
+    @mock.patch.object(logging, 'getLogger')
     @mock.patch.object(redis.StrictRedis, 'from_url')
-    def response(self, connection_init):
+    def response(self, connection_init, get_logger):
         connection_init.return_value = mock.Mock()
+        get_logger.return_value = mock.Mock()
 
         response = RedisTemplateResponse(request=None, template='')
 
@@ -21,13 +24,12 @@ class TestRedisTemplateResponse:
         response.resolve_template(template=['foo'])
         response.connection.get.assert_called_with('foo')
 
-    @mock.patch.object(logger, 'error')
-    def test_it_logs_error_and_raises_404_if_no_body(self, log_error, response):
+    def test_it_logs_error_and_raises_404_if_no_body(self, response):
         response.connection.get.return_value = None
         with pytest.raises(Http404):
             response.resolve_template([''])
 
-        assert log_error.called
+        assert response.logger.error.called
 
     def test_it_checks_for_another_template_on_current_version(self, response):
         response.resolve_template(['app:current'])
@@ -37,7 +39,5 @@ class TestRedisTemplateResponse:
         assert not hasattr(RedisTemplateResponse, 'connection')
         assert response.connection
 
-    def test_it_pings_connection_on_init(self):
-        response = self.response()
-
+    def test_it_pings_connection_on_init(self, response):
         assert response.connection.ping.called